tests: move entitlement_factory to base conftest
We need to mock entitlements in other tests besides the
ones in the entitlement package
